Analytics refactor
